+++ CHANGES.md
@@ -1,4 +1,5 @@
 ## 1.2.0 (unreleased)
+- Fix delegating script factory for keyed objects
 - Allow manual IDE hints for delegation to the RW class (see [#101](https://github.com/klum-dsl/klum-ast/issues/101))
 - Collection factories can now include templates (see [#77](https://github.com/klum-dsl/klum-ast/issues/77))
 - Replaced the `@ReadOnly` annotation with a `@Field(FieldType.PROTECTED)` (see [#108](https://github.com/klum-dsl/klum-ast/issues/108))